requirements.txt
kafka-python
pydantic
pydantic-settings
python-dotenv
aioimaplib